+++ src/main/java/com/agilebank/model/account/AccountDto.java
@@ -5,10 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NonNull;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class AccountDto {
   @NonNull private String id;
-  @NonNull private Double balance;
+  @NonNull private BigDecimal balance;
   @NonNull private Currency currency;
 }
