+++ src/main/java/com/agilebank/model/transaction/TransactionDto.java
@@ -5,11 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NonNull;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class TransactionDto {
   @NonNull private String sourceAccountId;
   @NonNull private String targetAccountId;
-  @NonNull private Double amount;
+  @NonNull private BigDecimal amount;
   @NonNull private Currency currency;
 }
